rename class Object because `object` has become a reserved word which can not be used as class name in PHP 7.2
